+++ CHANGELOG.rst
@@ -6,6 +6,8 @@ Changelog for stdlib_utils
 
 - Added kwarg for is_queue_eventually_empty and is_queue_eventually_empty for timeout
 
+- Added put_object_into_queue_and_raise_error_if_eventually_still_empty helper function for unit testing
+
 
 0.3.1 (09-17-20)
 ------------------